Humanizer_Version_02/requirements.txt
Flask==3.1.2
protobuf==6.32.1
python-dotenv==1.1.1
Requests==2.32.5
sentence_transformers==5.0.0
spacy==3.8.7
google-genai